Renamed artifact store module for simplicity
