+++ mlte/artifact/artifact.py
@@ -16,10 +16,7 @@ from mlte.artifact.model import ArtifactHeaderModel, ArtifactModel
 from mlte.artifact.type import ArtifactType
 from mlte.context.context import Context
 from mlte.session.state import session
-from mlte.store.artifact.artifact_store import (
-    ArtifactStore,
-    ManagedArtifactSession,
-)
+from mlte.store.artifact.store import ArtifactStore, ManagedArtifactSession
 
 
 class Artifact(metaclass=abc.ABCMeta):
