+++ mlte/store/artifact/underlying/http.py
@@ -23,11 +23,8 @@ from mlte.context.model import (
     Version,
     VersionCreate,
 )
-from mlte.store.artifact.artifact_store import (
-    ArtifactStore,
-    ArtifactStoreSession,
-)
 from mlte.store.artifact.query import Query
+from mlte.store.artifact.store import ArtifactStore, ArtifactStoreSession
 from mlte.store.base import StoreURI
 from mlte.web.store.api.model import WriteArtifactRequest
