+++ mlte/web/store/api/dependencies.py
@@ -15,7 +15,7 @@ solution that involves manual context management with a global state object.
 from contextlib import contextmanager
 from typing import Generator
 
-from mlte.store.artifact.artifact_store import ArtifactStoreSession
+from mlte.store.artifact.store import ArtifactStoreSession
 from mlte.web.store.state import state
