+++ test/measurement/memory/test_local_process_memory_consumption.py
@@ -17,7 +17,7 @@ from mlte.measurement.memory import (
     MemoryStatistics,
 )
 from mlte.spec.condition import Condition
-from mlte.store.artifact.artifact_store import ArtifactStore
+from mlte.store.artifact.store import ArtifactStore
 from mlte.validation.result import Failure, Success
 
 from ...fixture.store import store_with_context  # noqa
