+++ test/report/test_artifact.py
@@ -11,7 +11,7 @@ from mlte.model.shared import ProblemType
 from mlte.negotiation.artifact import NegotiationCard
 from mlte.report.artifact import Report
 from mlte.report.model import SummaryDescriptor
-from mlte.store.artifact.artifact_store import ArtifactStore
+from mlte.store.artifact.store import ArtifactStore
 
 from ..fixture.store import store_with_context  # noqa
