+++ test/value/types/test_image.py
@@ -13,7 +13,7 @@ import requests
 
 from mlte.context.context import Context
 from mlte.evidence.metadata import EvidenceMetadata, Identifier
-from mlte.store.artifact.artifact_store import ArtifactStore
+from mlte.store.artifact.store import ArtifactStore
 from mlte.value.types.image import Image
 
 from ...fixture.store import store_with_context  # noqa
